adam/ufpbots/requirements.txt
matplotlib==3.0.2
numpy==1.15.4
opencv-python==4.0.0.21
Pillow==5.4.1
pyserial==3.4
pytk==0.0.1.4
